Refactor paths to capitalize 'SuggestedMusic' directory
Updated file paths and imports to reflect capitalization of 'SuggestedMusic' directory for consistency and clarity. Also added new RSVP component under 'butterfly-vision' template structure.
