+++ src/views/non-authenticated/templates/butterfly-vision/RSVP/RSVP.vue
@@ -1,0 +1,20 @@
+<script setup>
+
+</script>
+
+<template>
+  <div
+    class="event-handle w-[70%] rounded-lg flex flex-col items-center"
+  >
+    <h2 class="text-6xl font-gvibes font-bold gap-10 gradient-text">
+      {{ title }}
+    </h2>
+    <h4 class="music-subtitle relative text-2xl font-normal text-gray-600/70 text-center moments-title">
+      {{ subTitle }}
+    </h4>
+  </div>
+</template>
+
+<style scoped>
+
+</style>
